+++ crates/model/src/instruments/synthetic.rs
@@ -120,12 +120,13 @@ impl<'de> Deserialize<'de> for SyntheticInstrument {
 impl SyntheticInstrument {
     /// Creates a new [`SyntheticInstrument`] instance with correctness checking.
     ///
-    /// # Notes
-    ///
-    /// PyO3 requires a `Result` type for proper error handling and stacktrace printing in Python.
     /// # Errors
     ///
     /// Returns an error if any input validation fails.
+    ///
+    /// # Notes
+    ///
+    /// PyO3 requires a `Result` type for proper error handling and stacktrace printing in Python.
     pub fn new_checked(
         symbol: Symbol,
         price_precision: u8,